refactor(services): split OCPP debug services into their own module - Related to #920
Adding ocpp_debug_set_current pushed services.py to 32343 bytes, over the
repository's 30 KB limit.

Split along the obvious seam: every ocpp_debug_* service talks directly to
a connected charger, bypassing the planner and the anti-flap state
machine, and exists to answer a question about the charger rather than to
run the product. That is a different job from force_recalculation or
set_temporary_override, and it is now six services sharing one resolution
helper.

_get_coordinator is imported inside _resolve_connected_charger rather than
at module scope, since services.py imports these handlers to build its
registration map and a module-level import back would be circular. The
deferred-import pattern is already used elsewhere in services.py.

No behaviour change: all 11 services still register, all 3159 tests pass.
